project/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Werkzeug==2.3.6
schedule==1.2.0
